cf-json-to-bq/requirements.txt
functions-framework==3.3.0
google-cloud-storage==2.10.0
google-cloud-bigquery==3.12.0
